@@ -10,7 +10,7 @@
   <PropertyGroup>
     <Description>MicroZero之实体操作Api基础</Description>
     <AssemblyTitle>Agebull.MicroZero.ModelApi</AssemblyTitle>
-    <TargetFrameworks>netcoreapp2.2</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1</TargetFrameworks>
     <DebugType>portable</DebugType>
     <AssemblyName></AssemblyName>
     <PackageId>Agebull.MicroZero.ModelApi</PackageId>
@@ -31,7 +31,7 @@
     <RunPostBuildEvent>OnOutputUpdated</RunPostBuildEvent>
     <Copyright />
     <PackageReleaseNotes>ZeroApi模型支持</PackageReleaseNotes>
-    <Version>1.4.10.37</Version>
+    <Version>1.5.0.3-bate</Version>
     <RootNamespace />
   </PropertyGroup>
 
@@ -43,11 +43,11 @@
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp2.2|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.1|AnyCPU'">
     <DocumentationFile>Agebull.MicroZero.ModelApi.xml</DocumentationFile>
   </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp2.2|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp3.1|AnyCPU'">
     <DocumentationFile>Agebull.MicroZero.ModelApi.xml</DocumentationFile>
   </PropertyGroup>
 
