@@ -3,7 +3,7 @@
   <PropertyGroup>
     <Description>ClrZmq的MicroZero专用版本</Description>
     <!--monoandroid60;xamarinios10-->
-    <TargetFrameworks>netcoreapp2.2</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1</TargetFrameworks>
     <DebugType>portable</DebugType>
     
     <PackageProjectUrl>https://github.com/agebullhu/MicroZero</PackageProjectUrl>
@@ -19,7 +19,7 @@
     <Company>Agebull</Company>
     <PackageReleaseNotes>ClrZmq的MicroZero专用版本</PackageReleaseNotes>
     <Product>Agebull.MicroZero</Product>
-    <Version>1.4.10.37</Version>
+    <Version>1.5.0.2-bate</Version>
     <!--<Version>1.3.0-alpha</Version>-->
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -30,14 +30,14 @@
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <NoWarn>1701;1702;1705;1591</NoWarn>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp2.2|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp3.1|AnyCPU'">
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <NoWarn>1701;1702;1705;1591</NoWarn>
     <DocumentationFile>Agebull.MicroZero.ClrZMQ.xml</DocumentationFile>
     <DefineConstants>TRACE</DefineConstants>
   </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp2.2|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.1|AnyCPU'">
     <DefineConstants>TRACE;DEBUG</DefineConstants>
     <DocumentationFile>Agebull.MicroZero.ClrZMQ.xml</DocumentationFile>
   </PropertyGroup>
@@ -82,8 +82,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Agebull.Common.LogRecorder" Version="2.4.10.21" />
-    <PackageReference Include="Agebull.EntityModel.Core" Version="2.4.10.18" />
+    <PackageReference Include="Agebull.Common.LogRecorder" Version="3.1.0.2-bate" />
+    <PackageReference Include="Agebull.EntityModel.Core" Version="3.1.0.2-bate" />
   </ItemGroup>
 
   <ItemGroup>
