@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <UserSecretsId>3909b577-6324-4fb6-b394-f48db7e0ad07</UserSecretsId>
     <StartupObject></StartupObject>
     <ApplicationIcon />
@@ -22,11 +22,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Senparc.Weixin.Cache.Redis" Version="2.5.10" />
-    <PackageReference Include="Senparc.Weixin.MP.MVC" Version="7.2.16" />
+    <PackageReference Include="Senparc.Weixin.Cache.Redis" Version="2.7.104" />
+    <PackageReference Include="Senparc.Weixin.MP.MVC" Version="7.4.104" />
   </ItemGroup>
   <ItemGroup>
     <Compile Remove="AddIn\**" />
