@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <UserSecretsId>3909b577-6324-4fb6-b394-f48db7e0ad07</UserSecretsId>
     <StartupObject></StartupObject>
     <ApplicationIcon />
@@ -21,9 +21,6 @@
     <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" />
-  </ItemGroup>
   <ItemGroup>
     <Compile Remove="AddIn\**" />
     <Compile Remove="App\**" />
