@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RunPostBuildEvent>Always</RunPostBuildEvent>
     <ApplicationIcon />
     <OutputType>Exe</OutputType>
