@@ -10,33 +10,33 @@ namespace Agebull.MicroZero
         /// <summary>
         ///     连接到
         /// </summary>
-        public async Task<bool> HeartLeft(string station, string realName)
+        public Task<bool> HeartLeft(string station, string realName)
         {
-            return ZeroApplication.ZerCenterIsRun && await ByteCommand(ZeroByteCommand.HeartLeft, station, realName);
+            return !ZeroApplication.ZerCenterIsRun ? Task.FromResult(false): ByteCommand(ZeroByteCommand.HeartLeft, station, realName);
         }
 
         /// <summary>
         ///     连接到
         /// </summary>
-        public async Task<bool> HeartReady(string station, string realName)
+        public Task<bool> HeartReady(string station, string realName)
         {
-            return ZeroApplication.ZerCenterIsRun && await ByteCommand(ZeroByteCommand.HeartReady, station, realName);
+            return !ZeroApplication.ZerCenterIsRun ? Task.FromResult(false): ByteCommand(ZeroByteCommand.HeartReady, station, realName);
         }
 
         /// <summary>
         ///     连接到
         /// </summary>
-        public async Task<bool> HeartJoin(string station, string realName)
+        public Task<bool> HeartJoin(string station, string realName)
         {
-            return ZeroApplication.ZerCenterIsRun && await ByteCommand(ZeroByteCommand.HeartJoin, station, realName, ZeroApplication.Config.LocalIpAddress);
+            return !ZeroApplication.ZerCenterIsRun ? Task.FromResult(false): ByteCommand(ZeroByteCommand.HeartJoin, station, realName, ZeroApplication.Config.LocalIpAddress);
         }
 
         /// <summary>
         ///     连接到
         /// </summary>
-        public async Task<bool> Heartbeat(string station, string realName)
+        public Task<bool> Heartbeat(string station, string realName)
         {
-            return ZeroApplication.ZerCenterIsRun && await ByteCommand(ZeroByteCommand.HeartPitpat, station, realName);
+            return !ZeroApplication.ZerCenterIsRun ? Task.FromResult(false): ByteCommand(ZeroByteCommand.HeartPitpat, station, realName);
         }
 
     }
